added CoordinateCalculation(たぶんMerge commit)
